Second try with DT after package update!
1) I submitted an issue comment and Yihui Xie from RStudio and he was
able to update the package and include a way to expose the last clicked
row to shiny. (Yay!, I hope other people find that features useful) — I
still can’t display it as a character in the sidebar-panel, but I’m
close, once I figure out all of the observe/reactive/action statements.
2) When I do get this working, I will trash the jQuery callbacks I
currently have from the previous attempts as they will not be needed.
3) With the update I also had to add “DT::” to all of the
renderDataTable and dataTableOutput. functions in the mainPanel.R and
the server-output.R i guess to differentiate it from the (soon to be
depreciated identical functions in the shiny package)