Editable table
1) There is a non reactive version using GSE13 for now: it correctly
edits and returns the correct table!
1a) REMOVED the uiOutput modal in favor of the numeric input for the
column to match the iris example and I got a lot of errors with
subsetting the table, but the numericInput works well
1b) the table editing capabilities also needed to not have the
dataTableAjax/action1 variables in order to have the correct number of
row names and column numbers
1c) the reactiveValues() needed to be in the shinySever from the
sourcing to access the variables so it has been moved from the above
the server to in the server-reactive.R
1e) I need to figure out how to incorporate a reactive/input$GSE or
just clinicalInput() in the values.edit$table to truly be reactive
because now it downloads the GEO number twice during loading
2) added library(shinyBS) to the mainPanel so it avoids the grey screen
of “death” for not finding bsModal.
3) commented out the palette since it was affecting the strip chart so
I have the col = colors() which looks dreary for the expression plot
but its a placeholder for finding an appropriate color scheme for the
entire app
4) rearranged the editClinicalTable() to closely match the iris example