Append to Report Button v1
This is a work in progress… but my issue at the moment is having the
plots show up on the knitr. The code is moderately correct but I
believe that it is a cache error. I’ll still looking into escape
chapters for the ```{r} code chunks for statements like this: x = \"",
profiles(), “\”, where i just need x = profiles()