:chart_with_downwards_trend: TO-DO: format axis/colors for Survival plot
1) modified the formatDE.R script with duplicates for outputs/functions
for the survival analyses page
1a) the bs modal correctly opens
1b) the group and label boxes appear with only one of the correct
labels from the time/outcome selection -> but these won’t appear on the
KM curve and yet the actual X and Y axis will need to be editable (ie.
‘Survival’ and ‘Time’ so that an user could specify ‘Time(months)’ for
example
1c) for the survival ‘formatTableDE2’ is a basic copy of the previous
with the lengths based on the input$survTimeUI and the
input$survOutcomeUI or I added a number 3 to differentiate the reactive
function
2) trying to associate the colorsDE3() with the col param in the
ggsurv() - I can get that to work but the color is the sample for both
curves. If you set surv.col = col you can see the effect
3) edited the server.R script to include the call for col = colorsDE3()
4) added button and bsModal to ui.R for the survival page!
5) more to do, but very close for this open task and issue #29