Added some basic test functionality.
When user clicks on Add Report under Differential Expression Analysis,
this will produce an R script named /path-to-server.r/test/GSE###_test.R
which is then rendered to HTML in this directory.

The output (FC and p-value) from the HTML file of the graph should match
what is observed from within Shiny.

Note that generating the test script requires a Mac (or Linux) machine. The
file GEO-script-template.R is used as a template for the test script, while
Shiny copies this file and specifies some parameters (such as GSE = "GSE13")
at the top.