Removed ```{r} hooks
removing the R hooks still outputs the code to the ace editor
appropriately and can be run back in the console by copy/paste

Note: Formatting from the formatDE script which controls the color and
labels for the Diff. Expo. and Survival Plot has too many reactive
dependencies so it won’t be able be included in the reproducible code
aspect.

This commit is prep for ONLY having the Code on the first release of
the web app for publication until the report is configured!