Added button so user can download new dataset without refreshing page. Use a shinyTitle reactive to dynamically change shiny title to include GSE and GPL info.
Notes - GSE and GPL info may look better on the right hand side of the header; will have to modify navBarPage code for this. Sometimes you have to click twice on the button for for the GSE and GPL modal. I'm not sure
exactly why this is, but it may be related to clicking on the button when the server side is processing.